% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datapath.R
\name{get_datapath}
\alias{get_datapath}
\title{Fetch the project data path}
\usage{
get_datapath(project_path = "", data_folder = "Data", use_dotenv = FALSE)
}
\arguments{
\item{project_path}{A characted string specifying the project path.}

\item{data_folder}{A character string specifying the data folder.}

\item{use_dotenv}{A logical value indicating whether to use the .env file to
get the project_path. If TRUE, the function will use the value specified in
the .env file. If FALSE, it will use the value specified by the user when
calling the function.}
}
\value{
A character string representing the path to the data directory.
}
\description{
Convenience function, which returns the path to the data directory of a
project in the form "C:/Users/username/project_path/data_folder".
}
\examples{
# not run
# get_datapath(project_path = my_project, data_folder = "R")
}
