% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_filetype.R
\name{subset_filetype}
\alias{subset_filetype}
\title{Subset files by file type}
\usage{
subset_filetype(name_vector, extension = "csv")
}
\arguments{
\item{name_vector}{A character vector of file names.}

\item{extension}{A character string specifying the file type to subset by.}
}
\value{
A character vector of file names that match the specified file type.
}
\description{
A helper function to subset a vector of file names by their file type.
}
\examples{
file_names <- c("data1.csv", "data2.tsv", "data3.xlsx",
"data4.txt", "data5.rds", "data6.json")
subset_filetype(file_names, "csv")
}
